cleaning and expanding Document object capability
