+++ dirlin/__init__.py
@@ -1,8 +1,10 @@
 """Dirlin (Directory Handler), set of tools to help process local files."""
 
-__version__ = "0.1.3"
+__version__ = "0.2.3"
 
-from .src.base import FolderPath as FolderPath
-from .src.base import Document as Document
-from .src.base import Directory as Directory
-from .src.base import Path as Path
+from .src.base.folder import (
+    Folder,
+    Document,
+    Directory,
+    Path
+)
