--- dirlin/pipeline/data_quality/validation/__init__.py
@@ -1,1 +1,0 @@
-from dirlin.pipeline.data_quality.validation.manager import BaseValidation
