+++ dirlin/src/base/validation.py
@@ -4,6 +4,8 @@ from typing import Any, Callable
 
 import pandas as pd
 
+from dirlin.src.base.util import DirlinFormatter
+
 
 @dataclasses.dataclass
 class _ResultWrapper:
@@ -19,139 +21,6 @@ class _ResultWrapper:
 
     check_name: str
     """stores the name of the check that will be put in the deliverable"""
-
-
-class _DirlinFormatter:
-    """utility object used for string formatting and series formatting
-    """
-    @classmethod
-    def convert_field_name_to_python_readable(cls, name: str) -> str:
-        """function for cleaning a column name. Can add onto this to cover more edge cases in the future.
-
-        The function will allow you to transform a standard column `California Taxes` to a python
-        usable format of `california_taxes`
-
-        :param name: the column name or the string we want to format to make it Python friendly
-        :return: a cleaned column name
-        """
-        name = name.strip("-").lower().replace(" ", "_")
-        return name
-
-    # todo should be convert string to number, with int or float as the param
-    @classmethod
-    def convert_string_field_to_integer_field(
-            cls,
-            key_field: pd.Series,
-            balance_field: pd.Series | None = None
-    ) -> pd.Series:
-        """may be used for something like limits in coverages, but will parse and format a pd.Series of strings
-        in a pd.Series of integers
-
-        :param key_field: a pd.Series of strings that we want to convert to a series of integers
-        :param balance_field: when given, will use the signs on the column to normalize against. For example,
-        a negative number on the balance field will convert the key_field to a negative as well
-
-        :return: a formatted pd.Series of integers
-        """
-        # (1) First stab at converting the key column. Assuming simplest case.
-        try:
-            key_field = key_field.fillna(0)
-            key_field = key_field.astype(float).astype(int)
-        except ValueError:
-            key_field = key_field.astype(str).fillna("0")
-
-            # formatting the string to clean up bad values
-            key_field = key_field.str.strip()
-            key_field = key_field.str.rstrip("_")
-            key_field = key_field.str.lower().replace("nan", "0")
-            key_field = key_field.str.replace("$", "")  # handle $
-            key_field = key_field.str.replace(",", "")  # handles commas
-            key_field = key_field.str.replace("_", "")
-            key_field = key_field.str.replace(' ', "")  # handles spaces
-            key_field = key_field.str.replace('no', "")  # handles no
-            key_field = key_field.str.replace('none', "")  # handles none
-
-        # (2) Second stab after trying to clean the column with any special values
-        try:
-            key_field = key_field.astype(str).fillna("0")
-            key_field = key_field.astype(float).astype(int)
-        except ValueError:
-            key_field = pd.to_numeric(key_field, errors='coerce', downcast='integer')
-            key_field = key_field.fillna(0)
-            key_field = key_field.astype(float).astype(int)
-        except Exception as e:
-            # catch all for now to future-proof any errors we might see
-            raise e
-
-        # (3) the optional parameter to normalize against the balance field
-        if isinstance(balance_field, pd.Series):
-            key_field = pd.Series(
-                [abs(amount) if b >= 0 else abs(amount) * -1 for amount, b in zip(key_field, balance_field)]
-            )
-        return key_field
-
-    @classmethod
-    def flatten_dict_to_list(cls, one_to_many_param: dict) -> list[dict]:
-        """used when ONE parameter has MANY columns associated with it, this function will convert
-        a nested list inside a dictionary, into a list of dictionaries with a key-value pair of
-        `parameter: column_name`
-
-        FROM one_to_many_param TO list of dictionaries
-        FROM `{A: [1, 2, 3], B: [A, B, C]}` TO `[{A: 1, B: A}, {A: 2, B: B}, {A: 3, B: C}]`
-
-        """
-        flattened_params_combo = [
-            dict(zip(one_to_many_param.keys(), values)) for values in zip(*one_to_many_param.values())
-        ]
-        if not flattened_params_combo:
-            flattened_params_combo = [dict()]
-        return flattened_params_combo
-
-    @classmethod
-    def format_args_reference_names(cls, arg_pair: dict[str, str], prefix: str | int | None = None) -> str:
-        """need this function in order to accept the argument from `map_function_to_args`.
-        This function allows us to reference the different arguments when we use the `run_validation` function.
-
-        Having a reference point allows us to create different variations of a final deliverable.
-        For example, we want to show check_a_v1: 100 errors, check_a_v2: 10 errors. This function creates the
-        `check_a_v1` and `check_a_v2` reference names.
-
-        :param arg_pair: the argument set we are going to use for a given function, likely comes from the
-        `map_function_to_args` function.
-
-        :return: a formatted reference name
-        """
-        cleaned_string = "_".join(
-            (cls.convert_field_name_to_python_readable(column_name) for column_name in arg_pair.values())
-        )
-        if prefix is not None:
-            cleaned_string = f"instance_{prefix}_{cleaned_string}"
-        return cleaned_string
-
-    @classmethod
-    def convert_string_field_to_percentage_field(cls, percentage_field: pd.Series) -> pd.Series:
-        """formats a percentage column with two decimal places and a percentage symbol
-
-        :param percentage_field: the string field with the percentage values
-        """
-        percentage_field = percentage_field.fillna("0").str.rstrip('%').astype(float)
-        return pd.Series([f"{round(float(value)): .2f}" for value in percentage_field])
-
-    @classmethod
-    def convert_percentage_field_to_float_field(cls, percentage_field: pd.Series) -> pd.Series:
-        """formats a percentage column into a float column
-
-        """
-        percentage_field = percentage_field.fillna("0").str.rstrip('%').astype(float) / 100  # todo conditional on val
-        return pd.Series(percentage_field)
-
-    @classmethod
-    def format_zip_fields(cls, zip_field: pd.Series) -> pd.Series:
-        """retains the leading 0 in a zip code string field
-
-        :param zip_field: the field with the Zip Code values
-        """
-        return zip_field.astype(str).str.extract('(\d+)', expand=False).str.zfill(5)
 
 
 class _BaseValidationVerifier:
@@ -329,7 +198,7 @@ class BaseValidation:
     """
 
     _validator: _BaseValidationVerifier = _BaseValidationVerifier
-    _formatter: _DirlinFormatter = _DirlinFormatter()
+    _formatter: DirlinFormatter = DirlinFormatter()
 
     @classmethod
     def run_validation(cls, df: pd.DataFrame) -> dict[str, dict]:
@@ -408,7 +277,7 @@ class BaseValidation:
         # We need this because arg list only gives you the field name and doesn't tie it to an existing DF
         deliverable = []
         new_keys = {
-            cls._formatter.format_args_reference_names(args_pair, "series"): {
+            cls._formatter.convert_dict_to_ref_names(args_pair, "series"): {
                 param: df[column] for param, column in args_pair.items()
             } for args_pair in args_list
         }  # this should look like `check_name`: `{param: pd.Series}` aka args_pair
@@ -449,7 +318,7 @@ class BaseValidation:
 
             # We need to iterate through these new results we just received
             results = [function(**args) for args in list_of_temp_args]
-            check_name = cls._formatter.format_args_reference_names(args_pair)
+            check_name = cls._formatter.convert_dict_to_ref_names(args_pair)
             deliverable.append(_ResultWrapper(result=pd.Series(results), check_name=check_name))
         return deliverable
 
@@ -488,7 +357,7 @@ class BaseValidation:
             # NOTE: realized that if I want to reference the params being used in the final deliverable,
             # NOTE cont: I need to be able to create a dict instead of a list to so that I can reference
             # NOTE cont: the changed parameter names
-            flat_one_to_many = cls._formatter.flatten_dict_to_list(one_to_many)
+            flat_one_to_many = cls._formatter.convert_dict_to_records(one_to_many)
             arg_sets: list[dict] = [one_to_one | one_to_many_args for one_to_many_args in flat_one_to_many]  # the args
 
             # 2025.03.21 previous way added too much complexity. Fixed to keep it clean.
