+++ test/test_pipelines/test_pipeline.py
@@ -2,7 +2,7 @@ import pandas as pd
 import pytest
 
 from dirlin import FolderPath
-from dirlin.pipeline import Pipeline, Report, Check, Validation
+from dirlin.src.pipeline import Pipeline, Report, Check, Validation
 
 
 class TestPipeLine:
